fix: add authentication form hook
