+++ packages/app/src/features/student/hooks/useAuthenticationForm.tsx
@@ -1,0 +1,37 @@
+import { useRouter } from 'next/router'
+import { useForm } from 'react-hook-form'
+import usePostAuthenticationFormMutation from '@features/student/queries/usePostAuthenticationFormMutation'
+import {
+  AuthenticationFormDto,
+  AuthenticationFormDtoSchema,
+} from '@features/student/dtos/form/AuthenticationFormDto'
+import ResToAuthenticationForm from '@features/student/lib/ResToAuthenticationForm'
+import { zodResolver } from '@hookform/resolvers/zod'
+import formToAuthenticationFormReq from '@features/student/lib/formToAuthenticationFormReq'
+import { AuthenticationFormResDto } from '@features/student/dtos/res/AuthenticationFromResDto'
+
+interface Props {
+  data?: AuthenticationFormResDto
+}
+
+const useAuthenticationForm = ({ data }: Props) => {
+  const router = useRouter()
+  const { mutate } = usePostAuthenticationFormMutation({
+    onSuccess: () => {
+      router.push('/')
+    },
+  })
+
+  const methods = useForm<AuthenticationFormDto>({
+    values: data ? ResToAuthenticationForm(data) : undefined,
+    resolver: zodResolver(AuthenticationFormDtoSchema),
+  })
+
+  const onSubmit = methods.handleSubmit(async (data) => {
+    mutate({ uuid: '', data: formToAuthenticationFormReq(data) })
+  })
+
+  return { onSubmit, methods }
+}
+
+export default useAuthenticationForm
